+++ lib/cobrato/entities/regress_cnab.rb
@@ -1,10 +1,6 @@
 module Cobrato
   module Entities
-    class RegressCnab < Base
-      attribute :id,               Integer
-      attribute :charge_config_id, Integer
-      attribute :status,           String
-      attribute :report,           Hash
+    class RegressCnab < Cnab
     end
   end
 end