+++ lib/cobrato/resources/regress_cnab.rb
@@ -5,6 +5,11 @@ module Cobrato
     class RegressCnab < Base
       crud :create, :show, :list, :destroy
 
+      def create(params)
+        deprecate(params, %w[charge_config_id])
+        super
+      end
+
       def file(id)
         http.get("#{resource_base_path}/#{id}/file") do |response|
           respond_with_openstruct(response)